store/mockstore: move mock storage to standalone package
